Add files via upload
This was my A level project so dont shout at me if it kinda sucks. it is trained on a csv of housing prices in sydney and then you can put in features you want your house to have and it will give you an estimated price. Id love to use this as a template for other projects if you know how to improve this please get in contact :3